ref: move wizarding into the module
