+++ src/wizarding/index.ts
@@ -1,3 +1,4 @@
 export * from './wizard.js'
 export * from './event.js'
 export * from './dom.js'
+export * from './wizarding.js'
